smoketests/requirements.txt
boto3==1.24.34
botocore==1.27.34
pytest>=7.2.2
configparser==5.3.0
snowflake-connector-python>=3.0.3
giturlparse.py>=0.0.5
arnparse==0.0.2
slack_sdk==3.22.0
python-hcl2==4.3.1
